Switch to Zig for Windows cross-compilation
Replaced mingw-w64 cross-compilers with Zig in Makefile and Go build logic for both x86 and x64 targets. Updated C++ code to use correct types for process/thread IDs and standardized AgentMain signature to match Windows thread entry point requirements. Added -fno-sanitize=all and improved linker flags for Windows subsystem compatibility.

+++ AdaptixServer/extenders/beacon_agent/src_beacon/beacon/AgentInfo.cpp
@@ -29,8 +29,8 @@ AgentInfo::AgentInfo()
 	this->acp           = ApiWin->GetACP();
 	this->oemcp         = ApiWin->GetOEMCP();
 	this->gmt_offest    = GetGmtOffset();
-	this->pid           = (WORD) NtCurrentTeb()->ClientId.UniqueProcess;
-	this->tid           = (WORD)NtCurrentTeb()->ClientId.UniqueThread;
+	this->pid           = (WORD) (uintptr_t) NtCurrentTeb()->ClientId.UniqueProcess;
+	this->tid           = (WORD) (uintptr_t) NtCurrentTeb()->ClientId.UniqueThread;
 	this->elevated      = IsElevate();
 	this->arch64        = (sizeof(void*) != 4);
 	this->sys64         = this->arch64 || isWow64;
